pricing  - price consumption chunks (reorg)
